requirements.txt
pyinstaller==6.0.0
pillow==10.2.0
